+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project follows [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [1.0.6-beta.1] - 2026-09-02
+
+### Added
+
+- Added **Pushcut** as a notification provider.
+- Added **webhook support**.
+
+### Performance
+
+- Optimized code for improved performance and efficiency.
+
 ## [1.0.5] - 2026-09-01
 
 ### Added
@@ -79,6 +90,7 @@ and this project follows [Semantic Versioning](https://semver.org/).
 - Homebridge custom configuration UI.
 
 [Unreleased]: https://github.com/mrfatboy/homebridge-snapshot-sensors/compare/main...NewFeature
+[1.0.6-beta.1]: https://github.com/mrfatboy/homebridge-snapshot-sensors/releases/tag/v1.0.6-beta.1
 [1.0.5]: https://github.com/mrfatboy/homebridge-snapshot-sensors/releases/tag/v1.0.5
 [1.0.4]: https://github.com/mrfatboy/homebridge-snapshot-sensors/releases/tag/v1.0.4
 [1.0.3]: https://github.com/mrfatboy/homebridge-snapshot-sensors/releases/tag/v1.0.3